requirements.txt
openai>=1.0.0
pypdf
chromadb
streamlit
python-dotenv
